+++ packages/server/graphql/public/mutations/generateInsight.ts
@@ -1,7 +1,7 @@
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
 import toTeamMemberId from '../../../../client/utils/relay/toTeamMemberId'
 import getKysely from '../../../postgres/getKysely'
-import {getUserId} from '../../../utils/authorization'
+import {getUserId, isSuperUser} from '../../../utils/authorization'
 import publish from '../../../utils/publish'
 import standardError from '../../../utils/standardError'
 import {MutationResolvers} from '../resolverTypes'
@@ -18,7 +18,7 @@ const generateInsight: MutationResolvers['generateInsight'] = async (
   const teamMemberId = toTeamMemberId(teamId, viewerId)
   const teamMember = await dataLoader.get('teamMembers').loadNonNull(teamMemberId)
   const isLead = teamMember.isLead
-  if (!isLead) {
+  if (!isLead && !isSuperUser(authToken)) {
     return standardError(new Error('Only team leads can generate insights'), {userId: viewerId})
   }
   const hasInsightsFlag = await dataLoader